stabilize cuDNN JAX launcher serialization
